% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockr.dag-dependencies.R
\name{add_blockr.dag_deps}
\alias{add_blockr.dag_deps}
\title{blockr.dag dependencies utils}
\usage{
add_blockr.dag_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches blockr.dag dependencies to the given tag
}
