% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{blk_icon_name}
\alias{blk_icon_name}
\title{Get block icon name}
\usage{
blk_icon_name(category)
}
\arguments{
\item{category}{Block category (old or new format)}
}
\value{
Character string with FontAwesome icon name
}
\description{
Returns the FontAwesome icon name for a block category.
}
\keyword{internal}
