% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockr.dag-package.R
\docType{package}
\name{blockr.dag-package}
\alias{blockr.dag}
\alias{blockr.dag-package}
\title{blockr.dag: A Directed Acyclic Graph Extension for 'blockr'}
\description{
Building on the docking layout manager provided by 'blockr.dock', this provides an extension that allows for visualizing and manipulating a 'blockr' board using a DAG-based user interface powered by the 'g6R' graph visualisation HTML widget.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bristolmyerssquibb.github.io/blockr.dag/}
}

}
\author{
\strong{Maintainer}: David Granjon \email{david@cynkra.com}

Authors:
\itemize{
  \item Nicolas Bennett \email{nicolas@cynkra.com}
  \item Christoph Sax \email{christoph@cynkra.com}
}

Other contributors:
\itemize{
  \item  Bristol Myers Squibb [funder]
}

}
\keyword{internal}
