% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g6r.R
\name{compute_children_from_links}
\alias{compute_children_from_links}
\title{Compute children relationships from links}
\usage{
compute_children_from_links(links)
}
\arguments{
\item{links}{Board links}
}
\value{
Named list where each element is a character vector of child node IDs
}
\description{
Compute children relationships from links
}
\keyword{internal}
