% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g6r.R
\name{g6_edges_from_links}
\alias{g6_edges_from_links}
\alias{g6_nodes_from_blocks}
\alias{g6_combos_data_from_stacks}
\alias{g6_data_from_board}
\title{Create network data from board}
\usage{
g6_edges_from_links(links, blocks, rankdir = "TB")

g6_nodes_from_blocks(blocks, stacks, children = NULL, rankdir = "TB")

g6_combos_data_from_stacks(stacks)

g6_data_from_board(board, rankdir = "TB")
}
\arguments{
\item{links}{Board links.}

\item{blocks}{Board blocks.}

\item{stacks}{Board stacks.}

\item{children}{Named list of children node IDs (optional).}

\item{board}{Board object.}
}
\description{
Create network data from board
}
\keyword{internal}
