% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g6r.R
\name{get_children_from_links}
\alias{get_children_from_links}
\title{Get children relationships from links}
\usage{
get_children_from_links(links)
}
\arguments{
\item{links}{Board links}
}
\value{
Named list where each element is a list of child node IDs (for JSON array conversion)
}
\description{
Get children relationships from links
}
\keyword{internal}
