% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g6r.R
\name{get_pending_node_position}
\alias{get_pending_node_position}
\title{Get and clear pending position for added node}
\usage{
get_pending_node_position()
}
\value{
List with x and y coordinates, or NULL
}
\description{
Get and clear pending position for added node
}
\keyword{internal}
