% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g6r.R
\name{is_variadic_block}
\alias{is_variadic_block}
\title{Check if a block is variadic}
\usage{
is_variadic_block(block)
}
\arguments{
\item{block}{A block object}
}
\value{
Logical indicating if the block is variadic
}
\description{
A block is variadic if it has NA arity and no named inputs.
}
\keyword{internal}
