% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{css_modal_advanced}
\alias{css_modal_advanced}
\alias{css_block_selectize}
\alias{js_blk_selectize_render}
\title{Shared CSS for modal advanced options and compact styling}
\usage{
css_modal_advanced(advanced_id)

css_block_selectize()

js_blk_selectize_render()
}
\arguments{
\item{advanced_id}{The namespaced ID for the advanced options container}
}
\value{
HTML style tag with modal CSS
}
\description{
Shared CSS for modal advanced options and compact styling
}
\keyword{internal}
