% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g6r.R
\name{resolve_target_ports}
\alias{resolve_target_ports}
\title{Resolve target port IDs for links}
\usage{
resolve_target_ports(links, blocks)
}
\arguments{
\item{links}{Board links}

\item{blocks}{Board blocks}
}
\value{
Character vector of target port IDs
}
\description{
Resolve target port IDs for links
}
\keyword{internal}
