% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g6r.R
\name{set_pending_node_position}
\alias{set_pending_node_position}
\title{Set pending position for next added node}
\usage{
set_pending_node_position(position)
}
\arguments{
\item{position}{List with x and y coordinates}
}
\description{
Set pending position for next added node
}
\keyword{internal}
