% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{blk_icon_name}
\alias{blk_icon_name}
\title{Get block icon name}
\usage{
blk_icon_name(category)
}
\arguments{
\item{category}{Block category}
}
\description{
Returns the FontAwesome icon name for a block category.
This is aligned with blockr.ui::blk_icon() but returns just the name.
}
\keyword{internal}
