% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g6r.R
\name{create_block_ports}
\alias{create_block_ports}
\title{Create block ports for g6 node}
\usage{
create_block_ports(block, id, rankdir = "TB")
}
\arguments{
\item{block}{Block object.}

\item{id}{Block ID.}
}
\description{
Create block ports for g6 node
}
\keyword{internal}
