% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext.R
\name{new_dag_extension}
\alias{new_dag_extension}
\title{DAG extension}
\usage{
new_dag_extension(graph = NULL, ...)
}
\arguments{
\item{graph}{A \code{graph} object (or \code{NULL})}

\item{...}{Forwarded to \code{\link[blockr.dock:extension]{blockr.dock::new_dock_extension()}}}
}
\description{
Visualizes the DAG (directed acyclic graph) underlying a board and provides
UI elemtnts to manipulate the board,
}
