% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{new_graph}
\alias{new_graph}
\alias{is_graph}
\alias{graph_nodes}
\alias{graph_edges}
\alias{graph_combos}
\title{Graph object}
\usage{
new_graph(nodes = list(), edges = list(), combos = list())

is_graph(x)

graph_nodes(x)

graph_edges(x)

graph_combos(x)
}
\arguments{
\item{nodes}{Graph nodes (i.e. board blocks)}

\item{edges}{Graph edges (i.e. board links)}

\item{combos}{Node groups (i.e. board stacks)}

\item{x}{Object}
}
\description{
R list-based representation of a g6 graph object.
}
