% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{new_dag_stack}
\alias{new_dag_stack}
\alias{is_dag_stack}
\alias{stack_color<-}
\alias{as_dag_stack}
\title{DAG stack}
\usage{
new_dag_stack(..., color = blockr.dock::suggest_new_colors())

is_dag_stack(x)

stack_color(x) <- value

as_dag_stack(x, ...)
}
\arguments{
\item{...}{Forwarded to \code{\link[blockr.core:new_stack]{blockr.core::new_stack()}}}

\item{color}{String-valued color}

\item{x}{object}

\item{value}{Replacement value}
}
\description{
Extends \code{\link[blockr.core:new_stack]{blockr.core::new_stack()}} by adding a color attribute.
}
