% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{new_toolbar_item}
\alias{new_toolbar_item}
\alias{is_toolbar_item}
\alias{toolbar_items}
\title{Create a toolbar item}
\usage{
new_toolbar_item(id, icon, js, action = NULL, tooltip = NULL)

is_toolbar_item(x)

toolbar_items(x)
}
\arguments{
\item{id}{Unique identifier for the toolbar item}

\item{icon}{Name of an icon to show in the toolbar}

\item{js}{JavaScript code to execute when the entry is selected}

\item{action}{Action to perform when the entry is selected}

\item{tooltip}{Optional tooltip text for the entry.}

\item{x}{Object}
}
\description{
Utilities to create tooblbar items for working with the DAG.
}
